feat: position 관련 suspense 추가
