+++ src/features/user/private/contents/userProfileUpdate/UserInfoForm.tsx
@@ -7,7 +7,7 @@ import { useUserDetailInfo } from '@/features/user/private/service/getUserDetail
 import { useEffect } from 'react';
 import Input from '@/shared/ui/Input';
 import { NicknameControl } from '@/features/user/private/components/userProfileUpdate/form/NicknameControl';
-import PositionControl from '@/features/user/private/components/userProfileUpdate/form/PositionControl';
+import UpdatePositionControl from '@/features/user/private/components/userProfileUpdate/form/UpdatePositionControl';
 import UpdateTechStackControl from '@/features/user/private/components/userProfileUpdate/form/UpdateTechStackControl';
 import IntroductionControl from '@/features/user/private/components/userProfileUpdate/form/IntroductionControl';
 import { bigIntToString } from '@/shared/utils/stringUtils';
@@ -57,7 +57,7 @@ const UserInfoForm = () => {
     <>
       <Input id='email' label='이메일' required disabled defaultValue={email} />
       <NicknameControl />
-      <PositionControl />
+      <UpdatePositionControl />
       <UpdateTechStackControl />
       <IntroductionControl />
     </>
